requirements.txt
gradio
openai-whisper
translate
elevenlabs
python-dotenv
